@@ -31,9 +31,7 @@ class GreeterServiceStub:
 
 
 if __name__ == "__main__":
-    reference_config = ReferenceConfig.from_url(
-        "tri://127.0.0.1:50051/org.apache.dubbo.samples.data.Greeter"
-    )
+    reference_config = ReferenceConfig.from_url("tri://127.0.0.1:50051/org.apache.dubbo.samples.data.Greeter")
     dubbo_client = dubbo.Client(reference_config)
     stub = GreeterServiceStub(dubbo_client)
 
